feat: ajustes

+++ src/main/JacksonVeroneze.TemplateWebApi.Domain/ValueObjects/CoordinateValueObject.cs
@@ -6,16 +6,16 @@ namespace JacksonVeroneze.TemplateWebApi.Domain.ValueObjects;
 
 public class CoordinateValueObject : ValueObject
 {
-    public double Latitude { get; private set; }
+    public float Latitude { get; private set; }
 
-    public double Longitude { get; private set; }
+    public float Longitude { get; private set; }
 
     protected CoordinateValueObject()
     {
     }
 
     private CoordinateValueObject(
-        double latitude, double longitude)
+        float latitude, float longitude)
     {
         Latitude = latitude;
         Longitude = longitude;
@@ -26,7 +26,7 @@ public class CoordinateValueObject : ValueObject
         => $"Lat: {value.Latitude} - Lon: {value.Longitude}";
 
     public static IResult<CoordinateValueObject> Create(
-        double latitude, double longitude)
+        float latitude, float longitude)
     {
         if (latitude <= 0 || longitude <= 0)
         {
